requirements.txt
requests>=2.32.0
pandas>=2.0.0
numpy>=1.26.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
google-generativeai>=0.8.0
streamlit>=1.35.0
plotly>=5.20.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
joblib>=1.3.0
tiktoken>=0.7.0
